refactor: Extract run_pipeline processing stages into stages.rs module
This extracts the distinct processing stages of the pipeline
(decoding, framing, VAD, smoothing, speech segments, merging,
and non-voice inversion) into separate functions within a dedicated
stages module. This significantly reduces the cognitive complexity
and size of mod.rs, bringing it down to 298 lines (well below the
strict 500-line MAX_SOURCE_FILE_LOC limit) and satisfying Owlwatch.
